+++ src/edu/buffalo/cse116/MandelbrotSet.java
@@ -195,18 +195,18 @@ public class MandelbrotSet {
 	}
 	
 	
-	public double getXCordRect(int xRow, int xStart, int xWidth) {
+	public double getXCordRect(int xRow, double xStart, double xWidth) {
 		double xCord = -2.15 + (2.75 * xStart / 512) + ((2.75 * xWidth / 512) * xRow / 512);
 		return xCord;
 	}
 	
-	public double getYCordRect(int yCol, int yStart, int yWidth) {
+	public double getYCordRect(int yCol, double yStart, double yWidth) {
 		double yCord = -1.3 + (2.6 * yStart / 512) +  ((2.6 * yWidth / 512) * yCol / 512);
 		return yCord;
 	}
 	
 	
-	public int[][] rectangle(String i, String b, int e, int r, int er, int gg) {
+	public int[][] rectangle(String i, String b, double e, double r, double er, double gg) {
 		int a = Integer.parseInt(i);
 		int c = Integer.parseInt(b);
 		int[][] array = new int[512][512];
