+++ src/edu/buffalo/cse116/burningshipset.java
@@ -222,16 +222,16 @@ public class burningshipset {
 		return array;
 	}
 	
-	public double getXCordRect(int xRow, int xStart, int xWidth) {
+	public double getXCordRect(int xRow, double xStart, double xWidth) {
 		double xCord = -1.8 + (0.1 * xStart / 512) + ((0.1 * xWidth / 512) * xRow / 512);
 		return xCord;
 	}
 	
-	public double getYCordRect(int yCol, int yStart, int yWidth) {
+	public double getYCordRect(int yCol, double yStart, double yWidth) {
 		double yCord = -0.08 + (0.105 * yStart / 512) +  ((0.105 * yWidth / 512) * yCol / 512);
 		return yCord;
 	}
-	public int[][] rectangle(String i, String b, int e, int r, int er, int gg) {
+	public int[][] rectangle(String i, String b, double e, double r, double er, double gg) {
 		int a = Integer.parseInt(i);
 		int c = Integer.parseInt(b);
 		int[][] array = new int[512][512];
