+++ src/edu/buffalo/cse116/multibrotSet.java
@@ -205,7 +205,7 @@ public class multibrotSet {
 	 * @param xWidth
 	 * @return
 	 */
-	public double getXCordRect(int xRow, int xStart, int xWidth) {
+	public double getXCordRect(int xRow, double xStart, double xWidth) {
 		double xCord = -1 + (2.0 * xStart / 512) + ((2.0 * xWidth / 512) * xRow / 512);
 		return xCord;
 	}
@@ -217,7 +217,7 @@ public class multibrotSet {
 	 * @param yWidth
 	 * @return
 	 */
-	public double getYCordRect(int yCol, int yStart, int yWidth) {
+	public double getYCordRect(int yCol, double yStart, double yWidth) {
 		double yCord = -1.3 + (2.6 * yStart / 512) +  ((2.6 * yWidth / 512) * yCol / 512);
 		return yCord;
 	}
@@ -232,7 +232,7 @@ public class multibrotSet {
 	 * @param gg
 	 * @return
 	 */
-	public int[][] rectangle(String i, String b, int e, int r, int er, int gg) {
+	public int[][] rectangle(String i, String b, double e, double r, double er, double gg) {
 		int a = Integer.parseInt(i);
 		int c = Integer.parseInt(b);
 		int[][] array = new int[512][512];
